Product Update

+++ app/Http/Controllers/Dashboard/ImageController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers\Dashboard;
 
 use App\Http\Controllers\Controller;
+use App\Models\ProductImage;
+use App\Models\Products;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\File;
 
@@ -35,9 +37,21 @@ class ImageController extends Controller
     public function cancel(Request $request){
 
         $temp_path = public_path("uploads/temp/".$request->image);
+        $product_path = public_path("uploads/product/".$request->image);
 
-        if(File::exists($temp_path)){
-            File::delete($temp_path);
+        if(File::exists($temp_path) || File::exists($product_path)){
+            
+            if(File::exists($temp_path)){
+                File::delete($temp_path);
+            }elseif(File::exists($product_path)){
+
+                //delete image from db
+                ProductImage::where('image',$request->image)->delete();
+                
+                File::delete($product_path);
+               
+            }
+            
             return response()->json([
                'status' => 200,
                'message' => 'Image Cancelled Successfully',
